ccxzc
xzcxzcxzcxzcxzcxzc

+++ code/modules/spells/roguetown/acolyte/necra.dm
@@ -83,6 +83,13 @@
 	..()
 	return TRUE
 
+
+
+/*
+	DEATH'S DOOR
+*/
+
+
 /obj/effect/proc_holder/spell/invoked/deaths_door
 	name = "Death's Door"
 	range = 7
@@ -98,19 +105,20 @@
 	invocation_type = "shout"
 	associated_skill = /datum/skill/magic/holy
 	antimagic_allowed = TRUE
-	recharge_time = 180 SECONDS
+	recharge_time = 30 SECONDS
 	miracle = TRUE
 	devotion_cost = 30
 
+
 /obj/effect/proc_holder/spell/invoked/deaths_door/cast(list/targets, mob/living/user)
 	..()
+
 	var/turf/T = get_turf(targets[1])
 	if(!isopenturf(T))
 		to_chat(user, span_warning("The targeted location is blocked. I cannot open a doorway here."))
 		return FALSE
 	for (var/obj/structure/underworld_portal/e_portal in user.contents) // checks if the portal exists, and shits them out
 		if(istype(e_portal))
-			e_portal.dispelled = FALSE //we are recasting after dispelling, its safe to set this as false.
 			e_portal.spitout_mob(user, T)
 			return TRUE
 	if(!locate(/obj/structure/underworld_portal) in T)
@@ -129,56 +137,70 @@
 	anchored = TRUE
 	density = FALSE
 	var/mob/living/caster // stores the caster. obviously.
-	var/list/trapped = list()// stores the trapped.
+	var/mob/living/trapped // stores the trapped.
 	var/time_id
-	var/dispelled = FALSE //Safety check
+
 
 /obj/structure/underworld_portal/examine(mob/living/carbon/user)
 	. = ..()
+
 	if(user.mob_biotypes & MOB_UNDEAD)
 		. += "A temporary gateway to the underworld. [span_warning("Faintly, you can see clutching fingers in the dark, reaching for you. If you go through, you won't come back.")]"
 	else
 		. += "A temporary gateway to the underworld. You can hear faint whispers through it. [span_warning("It might be possible to step through.")]"
 
+	. += "[span_notice("As the caster, click on GRAB to store the portal, provided there are souls inside. Use HARM to destroy the portal.")]"
+	if(trapped)
+		. += "[span_notice("Right-click on the portal to pull trapped souls out.")]"
+
+
 /obj/structure/underworld_portal/attack_hand(mob/living/carbon/user, list/modifiers)
-	..()
 	if(user == caster)
-		for (var/thing in contents)
-			if (istype(thing, /mob/living/carbon))
+		var/mob/living/carbon/victim = locate(/mob/living/carbon) in contents
+		if(user.used_intent.type == INTENT_GRAB)
+			if(victim)
 				caster.contents.Add(src)
-				dispelled = TRUE
 				user.visible_message(
 					span_revenwarning("[user] dispels the doorway with a touch."),
-					span_purple("I close the gateway. Opening it again will release whatever is inside.")
+					span_purple("I close the gateway.")
 					)
 				return TRUE
-		qdel(src)
-		return TRUE
+
+		if(user.used_intent.type == INTENT_HARM)
+			if(victim)
+				to_chat(user, span_warning("There are still souls trapped inside!"))
+				return FALSE
+			qdel(src)
+			return TRUE
+		return FALSE
+
 	if(!do_after(user, 2 SECONDS, src))
-		return
+		return FALSE
 	gobble_mob(user, caster)
-	return TRUE
+
+	..()
 
 
 /obj/structure/underworld_portal/Destroy()
-	if(dispelled == FALSE)//Only do this if we DON'T close it ourselves,that means something ELSE -FUNNY- happend.
-		visible_message(span_revenwarning("The portal collapses with an angry hiss."))
-		if(trapped)
-			for(var/mob/living/inportal in trapped)
-				spitout_mob(inportal, src.loc)
+	visible_message(span_revenwarning("The portal collapses with an angry hiss."))
+	spitout_mob(caster, loc)
+
 	..()
 
 /obj/structure/underworld_portal/attack_right(mob/living/carbon/user, list/modifiers)
 	..()
-	if(user == caster)
-		if(trapped)
-			for(var/mob/living/inportal in trapped)
-				spitout_mob(inportal)
-				user.visible_message(
-							span_revenwarning("[user] gestures thier hand at the gateway to expel what is within."),
-							span_purple("I gesture at the gateway to release whatever is inside.")
-							)
-			return TRUE
+
+	if(user != caster)
+		return FALSE
+	spitout_mob(user, loc)
+	user.visible_message(
+				span_revenwarning("[user] gestures their hand at the gateway to expel what is within."),
+				span_purple("I gesture at the gateway to release whatever is inside.")
+			)
+	qdel(src)
+
+	return TRUE
+
 
 /obj/structure/underworld_portal/MouseDrop_T(atom/movable/O, mob/living/user)
 	if(!isliving(O))
@@ -189,11 +211,15 @@
 		return
 	if(!do_after_mob(user, O, 5 SECONDS))
 		return
-	gobble_mob(O)
+	if(O == caster)
+		return
+	gobble_mob(O, user)
 	user.visible_message(
 		span_warning("[user] forces [O] into the portal!")
 	)
+
 	return TRUE
+
 
 /obj/structure/underworld_portal/proc/gobble_mob(mob/living/carbon/user, mob/living/carbon/caster)
 	if(user.mob_biotypes & MOB_UNDEAD)
@@ -210,39 +236,54 @@
 		span_revenwarning("[user] slips through the portal. Silence follows."),
 		span_purple("I touch the doorway. I slip through, and the world is silent and dark. I hear the distant rattle of a passing carriage.")
 		)
-	trapped += user
-	user.forceMove(src)
-	ADD_TRAIT(user, TRAIT_BLOODLOSS_IMMUNE, STATUS_EFFECT_TRAIT)
-	ADD_TRAIT(user, TRAIT_NOBREATH, STATUS_EFFECT_TRAIT)
-	user.add_client_colour(/datum/client_colour/monochrome)
+
+	if(user.mind)
+		if(trapped)
+			to_chat(user, span_warning("There is already a soul trapped inside!"))
+			return FALSE
+		user.forceMove(src)
+		ADD_TRAIT(user, TRAIT_BLOODLOSS_IMMUNE, STATUS_EFFECT_TRAIT)
+		ADD_TRAIT(user, TRAIT_NOBREATH, STATUS_EFFECT_TRAIT)
+		user.add_client_colour(/datum/client_colour/monochrome)
+		trapped = user
+	contents.Add(user)
 	time_id = addtimer(CALLBACK(src, PROC_REF(spitout_mob), user, null), 5 MINUTES, TIMER_UNIQUE | TIMER_OVERRIDE | TIMER_STOPPABLE) // 5 mins timer else its spitting you out where the necran is.
+
 	return TRUE
+
 
 /obj/structure/underworld_portal/proc/spitout_mob(mob/living/carbon/user, turf/T)
-	if(!trapped)
-		return FALSE
-	if(src in user.contents)
-		forceMove(T)
+	if(loc == user)
+		forceMove(T ? T : user.loc)
+		user.contents.Remove(src)
 
-	if(dispelled == TRUE)//dispelled at the caster, this is the case of we do not recast out dispelled portal and its been five minutes.
-		user.forceMove(caster.loc)
-		dispelled = FALSE
-	else
-		user.forceMove(loc)
+	if(trapped)
+		trapped.forceMove(loc)
+		contents.Remove(trapped)
+		trapped.remove_client_colour(/datum/client_colour/monochrome)
+		REMOVE_TRAIT(trapped, TRAIT_BLOODLOSS_IMMUNE, STATUS_EFFECT_TRAIT)
+		REMOVE_TRAIT(trapped, TRAIT_NOBREATH, STATUS_EFFECT_TRAIT)
+		trapped = null
+
+		trapped.visible_message(
+			span_revenwarning("[trapped] slips out from the whispering portal. Shadow roils off their form like smoke."),
+			span_purple("I am pulled from Necra's realm. Air fills my lungs, my heart starts beating- I live.")
+		)
+
+	for(var/mob/living/thing in contents)
+		if(istype(thing, /mob/living))
+			contents.Remove(thing)
+			thing.forceMove(loc)
+
 	if(time_id)
 		deltimer(time_id)
-	user.visible_message(
-		span_revenwarning("[user] slips out from the whispering portal. Shadow roils off their form like smoke."),
-		span_purple("I am pulled from Necra's realm. Air fills my lungs, my heart starts beating- I live.")
-		)
-	user.remove_client_colour(/datum/client_colour/monochrome)
-	REMOVE_TRAIT(user, TRAIT_BLOODLOSS_IMMUNE, STATUS_EFFECT_TRAIT)
-	REMOVE_TRAIT(user, TRAIT_NOBREATH, STATUS_EFFECT_TRAIT)
-	trapped -= user
+
 	return TRUE
+
 
 /obj/structure/underworld_portal/container_resist(mob/living/user)
 	..()
+
 	if(trapped != user)
 		return
 	var/resist_prob = user.STASTR * 2.5
@@ -251,81 +292,49 @@
 	spitout_mob(user)
 
 
-/obj/effect/proc_holder/spell/targeted/soulspeak
-	name = "Speak with Soul"
-	range = 5
-	overlay_state = "speakwithdead"
-	releasedrain = 30
-	recharge_time = 30 SECONDS
-	req_items = list(/obj/item/clothing/neck/roguetown/psicross)
-	max_targets = 0
-	cast_without_targets = TRUE
-	sound = 'sound/magic/churn.ogg'
-	associated_skill = /datum/skill/magic/holy
-	invocation = "She-Below brooks thee respite, be heard, wanderer."
-	invocation_type = "whisper" //can be none, whisper, emote and shout
-	miracle = TRUE
-	devotion_cost = 30
+/obj/effect/proc_holder/spell/targeted/speakwithdead
+    name = "Speak with Dead"
+    range = 5
+    overlay_state = "speakwithdead"
+    releasedrain = 30
+    recharge_time = 30 SECONDS
+    req_items = list(/obj/item/clothing/neck/roguetown/psicross)
+    max_targets = 0
+    cast_without_targets = TRUE
+    sound = 'sound/magic/churn.ogg'
+    associated_skill = /datum/skill/magic/holy
+    invocation = "The echoes of the departed stir, speak, O fallen one."
+    invocation_type = "whisper" //can be none, whisper, emote, and shout
+    miracle = TRUE
+    devotion_cost = 30
 
-/obj/effect/proc_holder/spell/targeted/soulspeak/cast(list/targets,mob/user = usr)
-	var/mob/living/carbon/spirit/capturedsoul = null
-	var/list/souloptions = list()
-	var/list/itemstore = list()
-	for(var/mob/living/carbon/spirit/S in GLOB.mob_list)
-		if(S.summoned)
-			continue
-		if(!S.client)
-			continue
-		souloptions += S.livingname
-	var/pickedsoul = input(user, "Which wandering soul shall I commune with?", "Available Souls") as null|anything in souloptions
-	if(!pickedsoul)
-		to_chat(user, span_warning("I was unable to commune with a soul."))
-		return
-	for(var/mob/living/carbon/spirit/P in GLOB.mob_list)
-		if(P.livingname == pickedsoul)
-			to_chat(P, "<font color='blue'>You feel yourself being pulled out of the Underworld.</font>")
-			sleep(2 SECONDS)
-			if(QDELETED(P) || P.summoned)
-				to_chat(user, "<font color='blue'>Your connection to the soul suddenly disappears!</font>")
-				return
-			capturedsoul = P
-			break
-	if(capturedsoul)
-		for(var/obj/item/I in capturedsoul.held_items) // this is still ass
-			capturedsoul.temporarilyRemoveItemFromInventory(I, force = TRUE)
-			itemstore += I.type
-			qdel(I)
-		capturedsoul.loc = user.loc
-		capturedsoul.summoned = TRUE
-		capturedsoul.beingmoved = TRUE
-		capturedsoul.invisibility = INVISIBILITY_OBSERVER
-		capturedsoul.status_flags |= GODMODE
-		capturedsoul.Stun(61 SECONDS)
-		capturedsoul.density = FALSE
-		addtimer(CALLBACK(src, PROC_REF(return_soul), user, capturedsoul, itemstore), 60 SECONDS)
-		addtimer(CALLBACK(src, PROC_REF(return_soul_warning), user, capturedsoul), 50 SECONDS)
-		to_chat(user, "<font color='blue'>I feel a cold chill run down my spine, a ghastly presence has arrived.</font>")
-		return ..()
+/obj/effect/proc_holder/spell/targeted/speakwithdead/cast(list/targets, mob/user = usr)
+    if(isliving(targets[1]) && targets[1].stat == DEAD)
+        var/mob/dead/target = targets[1]
+        return handle_dead_body(user, target)
+    else
+        to_chat(user, "<font color='red'>You cannot speak with the dead here.</font>")
+        return FALSE
 
-/obj/effect/proc_holder/spell/targeted/soulspeak/proc/return_soul_warning(mob/user, mob/living/carbon/spirit/soul)
-	if(!QDELETED(user))
-		to_chat(user, span_warning("The soul is being pulled away..."))
-	if(!QDELETED(soul))
-		to_chat(soul, span_warning("I'm starting to be pulled away..."))
+/proc/handle_dead_body(mob/user, mob/dead/target)
+    if(target.stat == DEAD && target.mind)
+        if(target.mind.active)
+            var/message = input(user, "You speak to the spirit of " + target.name + ". What will you say?", "Speak with the Dead")
 
-/obj/effect/proc_holder/spell/targeted/soulspeak/proc/return_soul(mob/user, mob/living/carbon/spirit/soul, list/itemstore)
-	to_chat(user, "<font color='blue'>The soul returns to the Underworld.</font>")
-	if(QDELETED(soul))
-		return
-	to_chat(soul, "<font color='blue'>You feel yourself being transported back to the Underworld.</font>")
-	soul.drop_all_held_items()
-	for(var/obj/effect/landmark/underworld/A in shuffle(GLOB.landmarks_list))
-		soul.loc = A.loc
-		for(var/I in itemstore)
-			soul.put_in_hands(new I())
-		break
-	soul.beingmoved = FALSE
-	soul.fully_heal(FALSE)
-	soul.invisibility = initial(soul.invisibility)
-	soul.status_flags &= ~GODMODE
-	soul.density = initial(soul.density)
+            if(message)
+                to_chat(target.mind, "<font color='blue'>" + user.name + " says: " + message + "</font>")
+                to_chat(user, "<font color='blue'>You speak to the spirit: " + message + "</font>")
+
+                var/spirit_message = input(target.mind, "The spirit responds to you: What will you ask?", "Spirit's Response")
+                if(spirit_message)
+                    to_chat(user, "<font color='blue'>The spirit replies: " + spirit_message + "</font>")
+                else
+                    to_chat(user, "<font color='blue'>The spirit chooses to remain silent.</font>")
+            else
+                to_chat(user, "<font color='blue'>You choose not to say anything.</font>")
+        else
+            to_chat(user, "<font color='blue'>The spirit has moved on and is no longer present.</font>")
+    else
+        to_chat(user, "<font color='blue'>No spirit is present to commune with.</font>")
+
+    return TRUE
